+++ include/utils.h
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   utils.h                                            :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: lvon-war <lvonwar@gmail.com>               +#+  +:+       +#+        */
+/*   By: lvon-war <lvon-war@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/03/11 12:14:25 by lvon-war          #+#    #+#             */
-/*   Updated: 2024/03/29 00:12:53 by lvon-war         ###   ########.fr       */
+/*   Updated: 2024/03/29 16:28:41 by lvon-war         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -34,6 +34,6 @@ void		free_texture(t_RGB *t);
 
 //utilsbis.c
 t_polygon	pointtopolygone(t_point a, t_point b, t_point c);
-int			degtorad(int degree);
+double		degtorad(double degree);
 
 #endif /*!CUBE_H */
